move session methods out of client_mfa module
