+++ crates/defguard_core/src/grpc/proxy/client_mfa.rs
@@ -45,7 +45,7 @@ use tonic::{Code, Status};
 use crate::{
     enterprise::{db::models::openid_provider::OpenIdProvider, is_business_license_active},
     events::{BidiRequestContext, BidiStreamEvent, BidiStreamEventType, DesktopClientMfaEvent},
-    grpc::{GatewayEvent, utils::parse_client_ip_agent},
+    grpc::{GatewayEvent, proxy::session::create_new_mfa_session, utils::parse_client_ip_agent},
 };
 
 const CLIENT_SESSION_TIMEOUT: u64 = 60 * 5; // 10 minutes
@@ -685,13 +685,15 @@ impl ClientMfaServer {
         let key = WireguardNetwork::genkey();
 
         // create new VPN client session
-        let vpn_client_session = self.create_new_mfa_session(
+        let vpn_client_session = create_new_mfa_session(
             &mut transaction,
             &location,
             &user,
             &device,
             method.into(),
             key.public.clone(),
+            self.wireguard_tx.clone(),
+            self.bidi_event_tx.clone(),
         )
         .await
             .map_err(|err| {
@@ -761,107 +763,6 @@ impl ClientMfaServer {
 
         Ok(response)
     }
-
-    /// Helper used to close all existing active sessions while creating a new MFA session
-    /// and send relevant gateway updates
-    async fn create_new_mfa_session(
-        &self,
-        conn: &mut PgConnection,
-        location: &WireguardNetwork<Id>,
-        user: &User<Id>,
-        device: &Device<Id>,
-        mfa_method: VpnClientMfaMethod,
-        preshared_key: String,
-    ) -> Result<VpnClientSession<Id>, Status> {
-        debug!(
-            "Creating new VPN session for device {device} of user {user} in location {location} after successful MFA authorization."
-        );
-
-        // find all active sessions for a given device and location
-        let active_sessions = VpnClientSession::get_all_active_device_sessions_in_location(&mut *conn, location.id, device.id).await
-            .map_err(|err| {
-                error!("Failed to fetch active VPN sessions for device {device} in location {location}: {err}");
-                Status::internal("unexpected error")
-            })?;
-        if !active_sessions.is_empty() {
-            info!(
-                "Found {} active sessions for device {device} in location {location}. Disconnecting them before creating a new MFA session",
-                active_sessions.len()
-            );
-        }
-
-        // disconnect all active sessions
-        for session in active_sessions {
-            debug!("Disconnecting previous active MFA VPN session {session:?}.");
-            self.disconnect_session(&mut *conn, session, location, user, device)
-                .await?;
-        }
-
-        // create new MFA session
-        let mut session =
-            VpnClientSession::new(location.id, user.id, device.id, None, Some(mfa_method));
-        session.preshared_key = Some(preshared_key);
-        session.save(conn).await.map_err(|err| {
-            error!("Failed to create new VPN client session for device {device} in location {location}: {err}");
-            Status::internal("unexpected error")
-        })
-    }
-
-    /// Update session state as disconnected and send relevant gateway update
-    async fn disconnect_session(
-        &self,
-        conn: &mut PgConnection,
-        mut session: VpnClientSession<Id>,
-        location: &WireguardNetwork<Id>,
-        user: &User<Id>,
-        device: &Device<Id>,
-    ) -> Result<(), Status> {
-        let is_connected = session.state == VpnClientSessionState::Connected;
-        let is_mfa_session = session.mfa_method.is_some();
-
-        // update session state in DB
-        let disconnect_timestamp = Utc::now().naive_utc();
-        session.disconnected_at = Some(disconnect_timestamp);
-        session.state = VpnClientSessionState::Disconnected;
-        session.save(&mut *conn).await.map_err(|err| {
-            error!("Failed to update VPN session {session:?}: {err}");
-            Status::internal("unexpected error")
-        })?;
-
-        // gateway update is only needed to remove peer for MFA sessions
-        // this is needed to remove peers for both Connected and New sessions
-        if is_mfa_session {
-            let gateway_event = GatewayEvent::MfaSessionDisconnected(location.id, device.clone());
-            self.wireguard_tx.send(gateway_event).map_err(|err| {
-                error!("Error sending WireGuard event: {err}");
-                Status::internal("unexpected error")
-            })?;
-        }
-
-        // only emit disconnect events if a session has actually been connected
-        if is_connected {
-            let context = BidiRequestContext {
-                timestamp: disconnect_timestamp,
-                user_id: user.id,
-                username: user.username.clone(),
-                ip: None,
-                device_name: format!("{device}"),
-            };
-            self.emit_event(BidiStreamEvent {
-                context,
-                event: BidiStreamEventType::DesktopClientMfa(Box::new(
-                    DesktopClientMfaEvent::Disconnected {
-                        location: location.clone(),
-                        device: device.clone(),
-                        is_mfa_session,
-                    },
-                )),
-            })
-            .map_err(Status::from)?;
-        }
-
-        Ok(())
-    }
 }
 
 #[cfg(test)]
@@ -893,7 +794,7 @@ mod tests {
     use super::{ClientLoginSession, ClientMfaServer};
     use crate::{
         events::{BidiStreamEvent, BidiStreamEventType, DesktopClientMfaEvent},
-        grpc::GatewayEvent,
+        grpc::{GatewayEvent, proxy::session::create_new_mfa_session},
     };
 
     const REPLACEMENT_MFA_PRESHARED_KEY: &str = "replacement-mfa-psk";
@@ -923,17 +824,18 @@ mod tests {
         let (server, mut event_rx, mut gateway_rx) = make_server(pool.clone());
         let mut conn = pool.acquire().await.expect("failed to acquire connection");
 
-        server
-            .create_new_mfa_session(
-                &mut conn,
-                &location,
-                &user,
-                &device,
-                VpnClientMfaMethod::Totp,
-                REPLACEMENT_MFA_PRESHARED_KEY.to_string(),
-            )
-            .await
-            .expect("should replace connected MFA session");
+        create_new_mfa_session(
+            &mut conn,
+            &location,
+            &user,
+            &device,
+            VpnClientMfaMethod::Totp,
+            REPLACEMENT_MFA_PRESHARED_KEY.to_string(),
+            server.wireguard_tx.clone(),
+            server.bidi_event_tx.clone(),
+        )
+        .await
+        .expect("should replace connected MFA session");
 
         let gateway_event = gateway_rx
             .try_recv()
@@ -998,17 +900,18 @@ mod tests {
         let (server, mut event_rx, mut gateway_rx) = make_server(pool.clone());
         let mut conn = pool.acquire().await.expect("failed to acquire connection");
 
-        server
-            .create_new_mfa_session(
-                &mut conn,
-                &location,
-                &user,
-                &device,
-                VpnClientMfaMethod::Totp,
-                REPLACEMENT_MFA_PRESHARED_KEY.to_string(),
-            )
-            .await
-            .expect("should replace new MFA session");
+        create_new_mfa_session(
+            &mut conn,
+            &location,
+            &user,
+            &device,
+            VpnClientMfaMethod::Totp,
+            REPLACEMENT_MFA_PRESHARED_KEY.to_string(),
+            server.wireguard_tx.clone(),
+            server.bidi_event_tx.clone(),
+        )
+        .await
+        .expect("should replace new MFA session");
 
         let gateway_event = gateway_rx
             .try_recv()
@@ -1057,17 +960,18 @@ mod tests {
         let (server, mut event_rx, mut gateway_rx) = make_server(pool.clone());
         let mut conn = pool.acquire().await.expect("failed to acquire connection");
 
-        server
-            .create_new_mfa_session(
-                &mut conn,
-                &location,
-                &user,
-                &device,
-                VpnClientMfaMethod::Totp,
-                REPLACEMENT_MFA_PRESHARED_KEY.to_string(),
-            )
-            .await
-            .expect("should replace connected non-MFA session");
+        create_new_mfa_session(
+            &mut conn,
+            &location,
+            &user,
+            &device,
+            VpnClientMfaMethod::Totp,
+            REPLACEMENT_MFA_PRESHARED_KEY.to_string(),
+            server.wireguard_tx.clone(),
+            server.bidi_event_tx.clone(),
+        )
+        .await
+        .expect("should replace connected non-MFA session");
 
         assert!(matches!(
             gateway_rx.try_recv(),
@@ -1197,17 +1101,18 @@ mod tests {
             .await
             .expect("failed to acquire database connection");
 
-        let new_session = server
-            .create_new_mfa_session(
-                &mut conn,
-                &location,
-                &user,
-                &device,
-                VpnClientMfaMethod::Totp,
-                NEW_MFA_PRESHARED_KEY.to_string(),
-            )
-            .await
-            .expect("failed to create replacement MFA session");
+        let new_session = create_new_mfa_session(
+            &mut conn,
+            &location,
+            &user,
+            &device,
+            VpnClientMfaMethod::Totp,
+            NEW_MFA_PRESHARED_KEY.to_string(),
+            server.wireguard_tx.clone(),
+            server.bidi_event_tx.clone(),
+        )
+        .await
+        .expect("failed to create replacement MFA session");
 
         let previous_session = VpnClientSession::find_by_id(&pool, previous_session.id)
             .await
