+++ crates/defguard_core/src/grpc/proxy/mod.rs
@@ -1,1 +1,2 @@
 pub mod client_mfa;
+pub(crate) mod session;
